@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+10
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=2.15.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -41,9 +41,11 @@ dependencies:
   in_app_review: ^2.0.4
   font_awesome_flutter: ^10.1.0
   url_launcher: ^6.1.5
-  get: ^4.6.5
-  get_storage: ^2.0.3
   flutter_launcher_icons: ^0.10.0
+  dio: ^4.0.6
+  connectivity_plus: ^3.0.2
+  dartz: ^0.10.1
+  get_it: ^7.2.0
   
 
   
