@@ -23,47 +23,44 @@ class ExpandedItem extends StatelessWidget {
     return Builder(
       builder: (context) => InkWell(
         onTap: () => onTap(),
-        child: Obx(
-          () => Container(
-            decoration: BoxDecoration(
-              border: Border.all(
-                color: context.colors.onBackground,
-              ),
-              borderRadius: const BorderRadius.all(
-                Radius.circular(90),
-              ),
-              color: clickedNumber == tabBarController.currentIndex.value
-                  ? context.colors.onBackground
-                  : Colors.transparent,
+        child: Container(
+          decoration: BoxDecoration(
+            border: Border.all(
+              color: context.colors.onBackground,
             ),
-            padding: EdgeInsets.symmetric(
-              horizontal: context.width(0.01),
-              vertical: context.height(0.006),
+            borderRadius: const BorderRadius.all(
+              Radius.circular(90),
             ),
-            child: Row(
-              mainAxisSize: MainAxisSize.min,
-              children: [
-                Text(
-                  text,
-                  textAlign: TextAlign.center,
-                  maxLines: 1,
-                  style: TextStyle(
-                    color: clickedNumber == tabBarController.currentIndex.value
-                        ? Colors.white
-                        : context.primary,
-                  ),
+            color: clickedNumber == tabBarController.currentIndex
+                ? context.colors.onBackground
+                : Colors.transparent,
+          ),
+          padding: EdgeInsets.symmetric(
+            horizontal: context.width(0.01),
+            vertical: context.height(0.006),
+          ),
+          child: Row(
+            mainAxisSize: MainAxisSize.min,
+            children: [
+              Text(
+                text,
+                textAlign: TextAlign.center,
+                maxLines: 1,
+                style: TextStyle(
+                  color: clickedNumber == tabBarController.currentIndex
+                      ? Colors.white
+                      : context.primary,
                 ),
-                Obx(() => Text(
-                      "($amount)",
-                      style: TextStyle(
-                        color:
-                            clickedNumber == tabBarController.currentIndex.value
-                                ? Colors.white
-                                : context.primary,
-                      ),
-                    ))
-              ],
-            ),
+              ),
+              Text(
+                "($amount)",
+                style: TextStyle(
+                  color: clickedNumber == tabBarController.currentIndex
+                      ? Colors.white
+                      : context.primary,
+                ),
+              ),
+            ],
           ),
         ),
       ),
