@@ -1,13 +1,9 @@
-import 'package:get/state_manager.dart';
+// import '../../../../../core/theme/theme_service.dart';
 
-import '../../../../../core/theme/theme_service.dart';
+// class ThemeController {
+//   ThemeController._();
+//   static final instance = ThemeController._();
+//   bool switchValue = ThemeService.instance.isSavedDarkMode();
 
-
-
-class ThemeController extends GetxController {
-  ThemeController._();
-  static final instance = ThemeController._();
-  RxBool switchValue = RxBool(ThemeService.instance.isSavedDarkMode());
-
-  void change(bool value) => switchValue.value = value;
-}
+//   void change(bool value) => switchValue = value;
+// }
