@@ -1,9 +1,11 @@
 import 'package:flutter/material.dart';
 
+import '../../../../core/navigation/contract/base_navigation_service.dart';
 import '../../../../global/export/export.dart';
 
 class SearchResultView extends StatelessWidget {
   final searchWordController = SearchController.instance;
+  final navigator = getIt<NavigationServiceContract>.call();
 
   SearchResultView({Key? key}) : super(key: key);
 
@@ -18,30 +20,31 @@ class SearchResultView extends StatelessWidget {
 
   CustomAppBar get _buildAppBar => CustomAppBar(
         title: const Text("Result"),
-        onTap: () => Get.back(),
+        onTap: () => navigator.getBack(),
       );
 
-  Widget get _getData => Obx(
-        () {
-          switch (searchWordController.getWord.runtimeType) {
-            case ErrorModel:
-              final ErrorModel errorModel = searchWordController.getWord;
-              return MyErrorWidget(
-                errorModel: ErrorModel(
-                  title: errorModel.title,
-                  message: errorModel.message,
-                ),
-              );
+  Widget get _getData => Builder(
+        builder: (context) {
+          return WordWidget(
+            wordModel: searchWordController.getWord,
+            controller: searchWordController,
+          );
 
-            case Null:
-              return const LoadingWidget();
+          // switch (searchWordController.getWord.runtimeType) {
+          //   case ErrorModel:
+          //     final ErrorModel errorModel = searchWordController.getWord;
+          //     return MyErrorWidget(
+          //       errorModel: ErrorModel(
+          //         title: errorModel.title,
+          //         message: errorModel.message,
+          //       ),
+          //     );
 
-            default:
-              return WordWidget(
-                wordModel: searchWordController.getWord,
-                controller: searchWordController,
-              );
-          }
+          //   case Null:
+          //     return const LoadingWidget();
+
+          //   default:
+          // }
         },
       );
 }
