@@ -1,65 +1,58 @@
 export 'package:auto_size_text/auto_size_text.dart';
 export 'package:flutter/cupertino.dart';
 export 'package:flutter/services.dart';
+export 'package:flutter_tts/flutter_tts.dart';
 export 'package:font_awesome_flutter/font_awesome_flutter.dart';
-export 'package:get/get_navigation/get_navigation.dart';
-export 'package:get/get_state_manager/get_state_manager.dart';
-export 'package:get/route_manager.dart';
-export 'package:get/state_manager.dart';
-export 'package:get_storage/get_storage.dart';
 export 'package:hive_flutter/hive_flutter.dart';
 export 'package:shared_preferences/shared_preferences.dart';
-export 'package:flutter_tts/flutter_tts.dart';
 
 export '../../../core/remote/admob/ad_helper.dart';
 export '../../../core/remote/admob/generate_ad.dart';
-export '../../../core/remote/api/services/word_service.dart';
-export '../../feature/presentation/controller/tab_controller.dart';
-
-export '../../feature/presentation/page/home/controller/accent_controller.dart';
-export '../../feature/presentation/page/home/controller/text_controller.dart';
-export '../../feature/presentation/page/home/controller/theme_controller.dart';
-export '../../feature/presentation/page/recent/controller/recent_controller.dart';
-export '../../feature/presentation/page/search_result/search_controller.dart/search_controller.dart';
-export '../../feature/presentation/page/word_of_the_day/controller/word_of_the_day_controller.dart';
+export '../../feature/data/api/services/word_repository_imp.dart';
 export '../../core/components/custom_app_bar.dart';
 export '../../core/components/error_widget.dart';
 export '../../core/components/loading_widget.dart';
 export '../../core/components/my_blinking_button.dart';
 export '../../core/extension/context_extension.dart';
 export '../../core/helpers/rating/rating_helper.dart';
+export '../../core/helpers/text_to_speech/text_to_speech.dart';
 export '../../core/helpers/url_launcher/url_launcher_helper.dart';
 export '../../core/initialization/initialization.dart';
 export '../../core/model/error_model.dart';
-export '../../core/navigation/app_pages.dart';
 export '../../core/remote/admob/ad_helper.dart';
 export '../../core/remote/admob/banner_ad_widget.dart';
-export '../../core/remote/api/services/word_service.dart';
+export '../../feature/data/api/services/word_repository_imp.dart';
 export '../../core/theme/theme.dart';
 export '../../core/theme/theme_service.dart';
-export '../../feature/presentation/components/word_widget.dart';
-export '../constant/kasset.dart';
-export '../constant/kcolor.dart';
-export '../constant/knetwork.dart';
-export '../constant/kstrings.dart';
-export '../../feature/presentation/page/about_me/about_me_view.dart';
-export '../../feature/presentation/page/home/components/accent_picker_widget.dart';
-export '../../feature/presentation/page/home/components/theme_picker_widget.dart';
-export '../../feature/presentation/page/home/controller/text_controller.dart';
-export '../../feature/presentation/page/recent/components/recent_item.dart';
-export '../../feature/presentation/page/recent/recent_view.dart';
-export '../../feature/presentation/page/search_result/search_controller.dart/search_controller.dart';
-export '../../feature/presentation/page/search_result/search_result_view.dart';
-export '../../feature/presentation/page/word_of_the_day/word_of_the_day_view.dart';
-export '../../feature/presentation/components/custom_text_field.dart';
-export '../../feature/presentation/controller/base_word_controller.dart';
-export '../../feature/presentation/controller/hive_controller.dart';
-export '../../feature/presentation/controller/hive_controller.dart';
 export '../../feature/data/model/definition.dart';
 export '../../feature/data/model/license.dart';
 export '../../feature/data/model/meaning.dart';
 export '../../feature/data/model/phonetic.dart';
 export '../../feature/data/model/word_model.dart';
-
-export '../../core/helpers/text_to_speech/text_to_speech.dart';
+export '../../feature/presentation/components/custom_text_field.dart';
+export '../../feature/presentation/components/word_widget.dart';
+export '../../feature/presentation/controller/base_word_controller.dart';
+export '../../feature/presentation/controller/hive_controller.dart';
+export '../../feature/presentation/controller/hive_controller.dart';
+export '../../feature/presentation/controller/tab_controller.dart';
+export '../../feature/presentation/page/about_me/about_me_view.dart';
+export '../../feature/presentation/page/home/components/accent_picker_widget.dart';
+export '../../feature/presentation/page/home/components/theme_picker_widget.dart';
+export '../../feature/presentation/page/home/controller/accent_controller.dart';
+export '../../feature/presentation/page/home/controller/text_controller.dart';
+export '../../feature/presentation/page/home/controller/text_controller.dart';
+export '../../feature/presentation/page/home/controller/theme_controller.dart';
+export '../../feature/presentation/page/recent/components/recent_item.dart';
+export '../../feature/presentation/page/recent/controller/recent_controller.dart';
+export '../../feature/presentation/page/recent/recent_view.dart';
+export '../../feature/presentation/page/search_result/search_controller.dart/search_controller.dart';
+export '../../feature/presentation/page/search_result/search_controller.dart/search_controller.dart';
+export '../../feature/presentation/page/search_result/search_result_view.dart';
+export '../../feature/presentation/page/word_of_the_day/controller/word_of_the_day_controller.dart';
+export '../../feature/presentation/page/word_of_the_day/word_of_the_day_view.dart';
+export '../constant/kasset.dart';
+export '../constant/kcolor.dart';
+export '../constant/knetwork.dart';
+export '../constant/kstrings.dart';
+export '../../injection_container.dart';
 
